long: turn the value field into a long and adjust accordingly
This should make most long support just work

+++ tree.h
@@ -9,7 +9,7 @@ struct node
 #define NOEFF			2	/* No side effect.. not yet used */
 #define NORETURN		4	/* Top level return is not used */
 #define ISBOOL			8	/* Return value is boolean truth */
-    unsigned value;		/* Offset for a NAME fp offset for a LOCAL */
+    unsigned long value;	/* Offset for a NAME fp offset for a LOCAL */
     unsigned snum;		/* Name of symbol (for code generator) */
     unsigned val2;		/* For code generator */
 };
